+++ module/pom.xml
@@ -138,28 +138,6 @@
           </dependency>
         </dependencies>
       </plugin>
-<!--      <plugin>-->
-<!--        <artifactId>maven-resources-plugin</artifactId>-->
-<!--        <version>${mavenResourcesVersion}</version>-->
-<!--        <executions>-->
-<!--          <execution>-->
-<!--            <id>copy-munit-resources</id>-->
-<!--            <phase>process-test-resources</phase>-->
-<!--            <goals>-->
-<!--              <goal>copy-resources</goal>-->
-<!--            </goals>-->
-<!--            <configuration>-->
-<!--              <outputDirectory>${munit.output.directory}</outputDirectory>-->
-<!--              <resources>-->
-<!--                <resource>-->
-<!--                  <directory>${munit.input.directory}</directory>-->
-<!--                  <filtering>true</filtering>-->
-<!--                </resource>-->
-<!--              </resources>-->
-<!--            </configuration>-->
-<!--          </execution>-->
-<!--        </executions>-->
-<!--      </plugin>-->
     </plugins>
   </build>
