+++ module/src/test/java/org/mule/extension/spring/test/SpringModuleAcceptVersionTestCase.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.spring;
+package org.mule.extension.spring.test;
 
 import static java.util.Optional.ofNullable;
 import static org.apache.commons.lang3.StringUtils.EMPTY;
