+++ module/src/test/java/org/mule/extension/spring/test/api/AllureConstants.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.spring.api;
+package org.mule.extension.spring.test.api;
 
 public interface AllureConstants {
