+++ module/src/test/java/org/mule/extension/spring/test/lifecycle/SpringLifecycleObject.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.spring.lifecycle;
+package org.mule.extension.spring.test.lifecycle;
 
 import static java.util.Collections.emptyList;
