+++ module/src/test/java/org/mule/extension/spring/test/notifications/ProcessorNotificationStore.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.spring.notifications;
+package org.mule.extension.spring.test.notifications;
 
 import static org.mule.runtime.api.notification.MessageProcessorNotification.MESSAGE_PROCESSOR_PRE_INVOKE;
