+++ module/src/test/resources/beans.xml
@@ -6,7 +6,7 @@
         http://www.springframework.org/schema/context
         http://www.springframework.org/schema/context/spring-context-4.1.xsd">
 
-    <bean id="customer" class="org.mule.extension.spring.api.SpringConfigTestCase.Customer">
+    <bean id="customer" class="org.mule.extension.spring.test.api.SpringConfigTestCase.Customer">
         <property name="name" value="customer name"/>
     </bean>
 </beans>
