+++ module/src/test/resources/jndi/jndi-beans.xml
@@ -13,7 +13,7 @@
     <jee:jndi-lookup id="jndiOrange" jndi-name="fruit/orange" environment-ref="jndiEnv"/>
 
     <util:properties id="jndiEnv">
-        <prop key="java.naming.factory.initial">org.mule.extension.spring.jndi.TestContextFactory</prop>
+        <prop key="java.naming.factory.initial">org.mule.extension.spring.test.jndi.TestContextFactory</prop>
     </util:properties>
 
 </beans>
