+++ module/src/test/resources/wildcardScan/spring-scan-wildcards-config-beans.xml
@@ -7,7 +7,7 @@
         http://www.springframework.org/schema/context/spring-context-4.1.xsd">
 
     <context:annotation-config />
-    <context:component-scan base-package="org.mule.extension.spring.scan*" />
+    <context:component-scan base-package="org.mule.extension.spring.test.scan*" />
 
 
 </beans>
